+++ index.js
@@ -51,10 +51,8 @@ module.exports = {
     app.import('bower_components/flatpickr-calendar/dist/flatpickr.min.css');
     app.import('bower_components/flatpickr-calendar/src/l10n/ru.js');
 
-    app.import({
-      development: 'bower_components/devicejs/lib/device.js',
-      production:  'bower_components/devicejs/lib/device.min.js'
-    });
+    // Custom devicejs
+    app.import('vendor/devicejs/devicejs.js');
 
     // Custom script which fixes some jQuery 1.10.x+ AJAX bugs with code from newer jQuery.3.0.0-alpha1+compat version.
     app.import('vendor/jquery.3.0.0-alpha1+compat/jquery.ajaxreplacement.js');