+++ tests/unit/services/device-test.js
@@ -5,8 +5,7 @@ import destroyApp from '../../helpers/destroy-app';
 
 let application;
 let appInstance;
-let originalNavigatorUserAgent;
-let originalNavigatorAppName;
+let stubbedNavigatorUserAgent;
 
 let cleanCache = function(deviceService) {
   deviceService.set('_cache.orientation', null);
@@ -17,58 +16,22 @@ let cleanCache = function(deviceService) {
 };
 
 let stubDevice = function(desiredDevice) {
-  var stubbedNavigatorUserAgent;
-  var stubbedNavigatoAppName;
 
   switch (desiredDevice) {
     case 'iphone':
       stubbedNavigatorUserAgent = 'Mozilla/5.0 (iPhone; CPU iPhone OS 10_2_1 like Mac OS X) AppleWebKit/602.1.50 (KHTML, like Gecko) ' +
         'Version/10.0 Mobile/14A456 Safari/602.1';
-      stubbedNavigatoAppName = 'Netscape';
       break;
   }
-
-  if (stubbedNavigatorUserAgent) {
-    window.navigator.__defineGetter__('userAgent', function () {
-      return stubbedNavigatorUserAgent;
-    });
-  }
-
-  if (stubbedNavigatoAppName) {
-    window.navigator.__defineGetter__('appName', function () {
-      return stubbedNavigatoAppName;
-    });
-  }
-};
-
-let storeOriginalDevice = function() {
-  if (!originalNavigatorUserAgent) {
-    originalNavigatorUserAgent = window.navigator.userAgent;
-  }
-
-  if (!originalNavigatorAppName) {
-    originalNavigatorAppName =  window.navigator.appName;
-  }
-};
-
-let resotreOriginalDevice = function() {
-  window.navigator.__defineGetter__('userAgent', function () {
-    return originalNavigatorUserAgent;
-  });
-  window.navigator.__defineGetter__('appName', function () {
-    return originalNavigatorUserAgent;
-  });
 };
 
 moduleForComponent('service:device', 'Unit | Service | Device', {
   integration: true,
   beforeEach() {
-    storeOriginalDevice();
     application = startApp();
     appInstance = application.buildInstance();
   },
   afterEach() {
-    resotreOriginalDevice();
     destroyApp(appInstance);
     destroyApp(application);
   }
@@ -125,19 +88,12 @@ test('work method platform(false)', function(assert) {
 
   stubDevice('iphone');
 
-  let $testPage = this.$();
-  let $scriptContainer = $('<div>').attr('id', 'deviceJsScriptContainer');
-  $testPage.append($scriptContainer);
-
-  // Force devicejs to run it's initialization again & read stubbed device-related properties.
-  let $devicejsScript = $('<script>').attr('type', 'text/javascript').attr('src', '/assets/devicejs.script');
-  $scriptContainer.append($devicejsScript);
-
   let service = application.__container__.lookup('service:device');
+  service.changeUserAgent(stubbedNavigatorUserAgent);
+
   let platform = service.platform(false);
   assert.strictEqual(platform, 'ios');
-
-  $scriptContainer.remove();
+  service.changeUserAgent();
 });
 
 test('work method type(false)', function(assert) {
@@ -145,17 +101,12 @@ test('work method type(false)', function(assert) {
 
   stubDevice('iphone');
 
-  let $testPage = this.$();
-  let $scriptContainer = $('<div>').attr('id', 'deviceJsScriptContainer');
-  $testPage.append($scriptContainer);
-
-  // Force devicejs to run it's initialization again & read stubbed device-related properties.
-  let $devicejsScript = $('<script>').attr('type', 'text/javascript').attr('src', '/assets/devicejs.script');
-  $scriptContainer.append($devicejsScript);
-
   let service = application.__container__.lookup('service:device');
+  service.changeUserAgent(stubbedNavigatorUserAgent);
+
   let result = service.type(false);
   assert.strictEqual(result, 'phone');
+  service.changeUserAgent();
 });
 
 test('work method pathPrefixes(false)', function(assert) {
@@ -163,15 +114,10 @@ test('work method pathPrefixes(false)', function(assert) {
 
   stubDevice('iphone');
 
-  let $testPage = this.$();
-  let $scriptContainer = $('<div>').attr('id', 'deviceJsScriptContainer');
-  $testPage.append($scriptContainer);
-
-  // Force devicejs to run it's initialization again & read stubbed device-related properties.
-  let $devicejsScript = $('<script>').attr('type', 'text/javascript').attr('src', '/assets/devicejs.script');
-  $scriptContainer.append($devicejsScript);
-
   let service = application.__container__.lookup('service:device');
+  service.changeUserAgent(stubbedNavigatorUserAgent);
+
   let result = service.pathPrefixes(false);
   assert.strictEqual(Ember.$.trim(result), 'mobile');
+  service.changeUserAgent();
 });